move helper functions to _project
