+++ components/results/ProjectResults.vue
@@ -107,16 +107,6 @@
   import TableHeader from '~/components/results/TableHeader.vue';
   import { mapGetters, mapMutations } from 'vuex';
   import specieSets from '~/refex-sample/datasets.json';
-  import _ from 'lodash';
-  const inRange = (x, [min, max]) => {
-    return typeof x !== 'number' || (x - min) * (x - max) <= 0;
-  };
-
-  const createNumberList = str =>
-    str
-      .replace('-', ',')
-      .split(',')
-      .map(x => parseInt(x) || 'out of filter bounds');
 
   export default {
     components: {
@@ -240,11 +230,6 @@
           column: col_name,
           selectedItem: this.selectedItem,
         });
-      },
-      textFilter(fullText, inputText) {
-        const reg = new RegExp(inputText, 'gi');
-        const isMatch = reg.test(fullText);
-        if (inputText.length > 0 && isMatch) return fullText.replaceAll(reg);
       },
       setDataset() {
         if (this.dataset === 'humanFantom5') {
